Creacion CRUD Operador

+++ database/migrations/2024_08_21_003149_create_operadorxusuario_table.php
@@ -15,7 +15,7 @@ class CreateOperadorxusuarioTable extends Migration
     {
         Schema::create('operadorxusuario', function (Blueprint $table) {
             $table->integer('cod_opexusu');
-            $table->integer('id_operador');
+            $table->unsignedBigInteger('id_operador');
             $table->integer('id_usuario');
 
             $table->primary(['cod_opexusu', 'id_operador', 'id_usuario']);
