+++ database/migrations/2024_08_21_111836_create_laboratorio_ii_trimestre_table.php
@@ -19,7 +19,7 @@ class CreateLaboratorioIiTrimestreTable extends Migration
             $table->integer('cod_opexusu');
             $table->foreign('cod_opexusu')->references('cod_opexusu')->on('operadorxusuario');
 
-            $table->integer('id_operador');
+            $table->unsignedBigInteger('id_operador');
             $table->foreign('id_operador')->references('id_operador')->on('operador');
 
             $table->integer('id_usuario');
